Update set-up-phone-system-voicemail.md
@dstrome 

Nathan Bennett and I verified all steps and realized we were missing important step for the Integration between SfB Online and Exchange which I added to Step 6. 

There will be other changes as discussed in email thread to try to coordinate Teams and Skype for Business Server documentation, but for now this bit is a priority as its really required to get this working for Exchange Server customers. 

Can you prioritize this PR? Thanks.